Move jsonconverters to seperate namespace/folder
