+++ src/main/Plugins/Base/Options/StorePluginOptions.cs
@@ -1,5 +1,6 @@
 ﻿using PKISharp.WACS.Plugins.Interfaces;
 using PKISharp.WACS.Services;
+using PKISharp.WACS.Services.Serialization;
 using System;
 
 namespace PKISharp.WACS.Plugins.Base.Options
