+++ src/main/Plugins/Base/Options/TargetPluginOptions.cs
@@ -1,6 +1,6 @@
-﻿using Newtonsoft.Json;
-using PKISharp.WACS.Plugins.Interfaces;
+﻿using PKISharp.WACS.Plugins.Interfaces;
 using PKISharp.WACS.Services;
+using PKISharp.WACS.Services.Serialization;
 using System;
 
 namespace PKISharp.WACS.Plugins.Base.Options
