+++ src/main/Plugins/Base/Options/ValidationPluginOptions.cs
@@ -2,6 +2,7 @@
 using Newtonsoft.Json;
 using PKISharp.WACS.Plugins.Interfaces;
 using PKISharp.WACS.Services;
+using PKISharp.WACS.Services.Serialization;
 using System;
 
 namespace PKISharp.WACS.Plugins.Base.Options
