+++ src/main/Plugins/Base/OptionsFactories/PluginOptionsFactory.cs
@@ -1,5 +1,6 @@
 ﻿using PKISharp.WACS.Plugins.Interfaces;
 using PKISharp.WACS.Services;
+using PKISharp.WACS.Services.Serialization;
 using System;
 
 namespace PKISharp.WACS.Plugins.Base.Factories
