+++ src/main/Plugins/StorePlugins/CentralSsl/CentralSslOptions.cs
@@ -2,6 +2,7 @@
 using PKISharp.WACS.Plugins.Base;
 using PKISharp.WACS.Plugins.Base.Options;
 using PKISharp.WACS.Services;
+using PKISharp.WACS.Services.Serialization;
 
 namespace PKISharp.WACS.Plugins.StorePlugins
 {
