+++ src/plugin.validation.dns.route53/Route53Options.cs
@@ -2,6 +2,7 @@
 using PKISharp.WACS.Services;
 using PKISharp.WACS.Plugins.Base;
 using PKISharp.WACS.Plugins.Base.Options;
+using PKISharp.WACS.Services.Serialization;
 
 namespace PKISharp.WACS.Plugins.ValidationPlugins.Dns
 {
